Split keymaster credential routes
