+++ services/keymaster/server/src/keymaster-api.ts
@@ -16,20 +16,27 @@ import { InvalidParameterError } from '@didcid/common/errors';
 import config from './config.js';
 import { createAddressRouter } from './keymaster-address-router.js';
 import { createAddressTailRouter } from './keymaster-address-tail-router.js';
+import { createAgentRouter } from './keymaster-agent-router.js';
 import { createAssetRouter } from './keymaster-asset-router.js';
+import { createChallengeRouter } from './keymaster-challenge-router.js';
 import { createRequireAdminKey } from './keymaster-admin.js';
 import { createCoreRouter } from './keymaster-core-router.js';
-import { createCredentialRouter } from './keymaster-credential-router.js';
+import { createCredentialWorkflowRouter } from './keymaster-credential-workflow-router.js';
 import { createDidCommRouter } from './keymaster-didcomm-router.js';
 import { createDmailRouter } from './keymaster-dmail-router.js';
 import { createFileRouter } from './keymaster-file-router.js';
+import { createGroupRouter } from './keymaster-group-router.js';
 import { createIdentityRouter } from './keymaster-identity-router.js';
 import { createImageRouter } from './keymaster-image-router.js';
+import { createKeyRouter } from './keymaster-key-router.js';
 import { createLightningRouter } from './keymaster-lightning-router.js';
 import { createNostrRouter } from './keymaster-nostr-router.js';
 import { createNoticeRouter } from './keymaster-notice-router.js';
 import { createPollRouter } from './keymaster-poll-router.js';
 import { createPublicRouter } from './keymaster-public-router.js';
+import { createResponseRouter } from './keymaster-response-router.js';
+import { createSchemaRouter } from './keymaster-schema-router.js';
+import { createSchemaTemplateRouter } from './keymaster-schema-template-router.js';
 import type { CreateKeymasterRouterOptions } from './keymaster-router-types.js';
 import { createVaultRouter } from './keymaster-vault-router.js';
 import promClient from 'prom-client';
@@ -176,7 +183,14 @@ v1router.use(createDidCommRouter(routerOptions));
 v1router.use(createAddressTailRouter(routerOptions));
 v1router.use(createNostrRouter(routerOptions));
 v1router.use(createLightningRouter(routerOptions));
-v1router.use(createCredentialRouter(routerOptions));
+v1router.use(createChallengeRouter(routerOptions));
+v1router.use(createResponseRouter(routerOptions));
+v1router.use(createGroupRouter(routerOptions));
+v1router.use(createSchemaRouter(routerOptions));
+v1router.use(createAgentRouter(routerOptions));
+v1router.use(createCredentialWorkflowRouter(routerOptions));
+v1router.use(createKeyRouter(routerOptions));
+v1router.use(createSchemaTemplateRouter(routerOptions));
 v1router.use(createAssetRouter(routerOptions));
 v1router.use(createPollRouter(routerOptions));
 v1router.use(createImageRouter(routerOptions));
