+++ swaggerConf.js
@@ -50,7 +50,14 @@ const keymasterOptions = {
         'services/keymaster/server/src/keymaster-address-tail-router.ts',
         'services/keymaster/server/src/keymaster-nostr-router.ts',
         'services/keymaster/server/src/keymaster-lightning-router.ts',
-        'services/keymaster/server/src/keymaster-credential-router.ts',
+        'services/keymaster/server/src/keymaster-challenge-router.ts',
+        'services/keymaster/server/src/keymaster-response-router.ts',
+        'services/keymaster/server/src/keymaster-group-router.ts',
+        'services/keymaster/server/src/keymaster-schema-router.ts',
+        'services/keymaster/server/src/keymaster-agent-router.ts',
+        'services/keymaster/server/src/keymaster-credential-workflow-router.ts',
+        'services/keymaster/server/src/keymaster-key-router.ts',
+        'services/keymaster/server/src/keymaster-schema-template-router.ts',
         'services/keymaster/server/src/keymaster-asset-router.ts',
         'services/keymaster/server/src/keymaster-poll-router.ts',
         'services/keymaster/server/src/keymaster-image-router.ts',
